+++ scripts/legacyMacros.js
@@ -650,6 +650,8 @@ export {wrapsOfDyamak, wrapsOfDyamak0, wrapsOfDyamak1, wrapsOfDyamak2, crimsonMi
 export {robeOfUsefulItems, robeOfUsefulItemsSetupPatch, robeOfUsefulItemsPatch} from './macros/2014/items/clothing/robeOfUsefulItems.js';
 export {ruinsWakeD, ruinsWakeA, ruinsWakeE} from './macros/2014/items/weapons/ruinsWake.js';
 export {charlatansDie} from './macros/2014/items/trinket/charlatansDie.js';
+export {deckOfMiscellany, deckOfMiscellanySetup, miscellanyCard} from './macros/2014/items/trinket/deckOfMiscellany.js';
+export {deckOfWonder, deckOfWonderDestinyEffect, deckOfWonderIsolationEffect, deckOfWonderJusticeEffect, deckOfWonderVultureLostEffect} from './macros/2014/items/trinket/deckOfWonder.js';
 //BG3 Weapon Actions
 export {backbreaker, braceMelee, braceMeleeDamage, cleave, concussiveSmash, maimingStrike, disarmingStrike, flourish, heartstopper, lacerate, piercingStrike, piercingStrikeEffect, pommelStrike, prepare, rush, tenacity, topple, weakeningStrike, braceRanged, braceRangedDamage, hamstringShot, mobileShot, piercingShot} from './macros/2014/homebrew/bg3WeaponActions.js';
 //Other