+++ scripts/macros/2024/spells/summonFey.js
@@ -81,8 +81,17 @@ async function use({workflow}) {
     }
     updates.actor.items.push(moodFeatureData);
     let animation = itemUtils.getConfig(workflow.item, creatureType + 'Animation') ?? 'none';
+    let duration = 3600;
+    const uuid = workflow.item.flags.dnd5e?.cachedFor;
+    if (uuid) {
+        const activity = await fromUuid(uuid, {relative: workflow.actor});
+        
+        if (activity?.item === itemUtils.getItemByIdentifier(workflow.actor, 'feyReinforcements')) {
+            duration = 60;
+        }
+    }
     await Summons.spawn(sourceActor, updates, workflow.item, workflow.token, {
-        duration: 3600,
+        duration: duration,
         range: 90,
         animation,
         initiativeType: 'follows',